+++ notifiers_linuxfreebsd.go
@@ -33,8 +33,9 @@ func bannerNotify(n notification) error {
 	}
 
 	cmd := exec.Command("notify-send", n.title, n.message)
+	cmd.Stderr = os.Stderr
 	if err = cmd.Run(); err != nil {
-		return err
+		return fmt.Errorf("Banner: %s", err)
 	}
 
 	return nil
@@ -63,8 +64,9 @@ func speechNotify(n notification) error {
 	text := fmt.Sprintf("%s %s", n.title, n.message)
 
 	cmd := exec.Command("espeak", "-v", voice, "--", text)
+	cmd.Stderr = os.Stderr
 	if err = cmd.Run(); err != nil {
-		return err
+		return fmt.Errorf("Speech: %s", err)
 	}
 
 	return nil